Rainfall.tsx: extracted out bar chart to new component
