+++ src/components/Rainfall.tsx
@@ -1,27 +1,16 @@
-import { useEffect, useState } from "react";
+// do not render bar chart if no rainfall?
+
+import { useEffect} from "react";
 import { useAppSelector, useAppDispatch } from "../app/hooks";
 import type { RootState } from "../app/store";
 import { updateRainfall } from "../features/rainfall/rainfallSlice";
 import { fetchRainfall } from "../models/apiQueries";
 import "../css/Rainfall.css";
-import { Card } from "@mui/material";
-//bar
-import {
-  XYPlot,
-  XAxis,
-  YAxis,
-  VerticalGridLines,
-  HorizontalGridLines,
-  VerticalBarSeries,
-} from "react-vis";
+import { Card, CardContent } from "@mui/material";
+import { RainfallBarChart } from "./RainfallBarChart";
+
 
 export const Rainfall = () => {
-  //variables
-
-  //states
-  const [chartRainfallData, setChartRainfallData] = useState([
-    { x: "default", y: 1 },
-  ]);
 
   //redux states
   /**
@@ -32,15 +21,7 @@ export const Rainfall = () => {
   const currentLocation = useAppSelector((state: RootState) => {
     return state.location;
   });
-
-  /**
-   * Gets rainfall state state from redux store.
-   *
-   * @returns {number[]} Array representing rainfall values of the day from midnight to 11pm
-   */
-  const currentRainfall = useAppSelector((state: RootState) => {
-    return state.rainfall;
-  });
+  
   //variables
   const dispatch = useAppDispatch();
 
@@ -58,55 +39,17 @@ export const Rainfall = () => {
     currentRainfall();
   }, [currentLocation, dispatch]);
 
-  //update rainfall data from store for bar chart
-  useEffect(() => {
-    const labels = [
-      "00:00",
-      "01:00",
-      "02:00",
-      "03:00",
-      "04:00",
-      "05:00",
-      "06:00",
-      "07:00",
-      "08:00",
-      "09:00",
-      "10:00",
-      "11:00",
-      "12:00",
-      "13:00",
-      "14:00",
-      "15:00",
-      "16:00",
-      "17:00",
-      "18:00",
-      "19:00",
-      "20:00",
-      "21:00",
-      "22:00",
-      "24:00",
-    ];
-console.log("rerender2")
-    const processedData = currentRainfall.map((rainfall, index) => {
-      return { x: labels[index], y: rainfall };
-    });
-    setChartRainfallData(processedData);
-  }, [currentRainfall]);
+
 
   return (
     <section>
       <Card style={{ border: "none", boxShadow: "none" }}>
-        <XYPlot xType="ordinal" width={1200} height={250}>
-          <VerticalGridLines />
-          <HorizontalGridLines />
-          <XAxis />
-          <YAxis />
-          <VerticalBarSeries
-            barWidth={1}
-            className="vertical-bar-series-example"
-            data={chartRainfallData}
-          />
-        </XYPlot>
+      <CardContent
+        style={{ backgroundColor: "#18191a" }}
+        sx={{ width: "75rem", height: "15rem" }}
+      >
+       <RainfallBarChart></RainfallBarChart>
+       </CardContent>
       </Card>
     </section>
   );
